[#371] $SELECT does not stop evaluating tvexprs or exprs after the first true tvexpr is found in case of NUMOFLOW or INVDLRCVAL errors
Let us take the below M line as an example.

	if $select(1:1,1:1E47)  write "Hello "

While compiling this, the moment 1E47 is encountered, advancewindow() issues a stx_error(ERR_NUMOFLOW).
And sets TREF(director_token) to TK_ERROR. This causes the parse of the entire IF command arguments to
be abandoned (including the $SELECT etc.) and instead an error triple to be inserted (by ins_errtriple()).
This results in a runtime error if ever the IF command is encountered.

This does not adhere to the ANSI M Standard which says the $select() tvexprs should be evaluated from
left to right and the moment a true one is found, no more tvexprs or exprs to the right of it should be
evaluated. In this case, the moment 1:1 (tvexpr:expr) is found, the 1:1E47 should not be evaluated.

The way this is fixed is to introduce a new token type TK_NUMOFLOWLIT to indicate in advancewindow()
whenever we notice a numeric literal that has an overflow (like 1E47). TREF(director_token) is not
set to TK_ERROR in this case and that lets parsing continue fine even after this NUMOFLOW point.

Once advancewindow() returns, the caller expritem() checks for TK_NUMOFLOWLIT and in that case it
generates a dummy triple (literal_numoflow, a new literal that is set to an overflow number with exponent
EXPHI) to indicate that was the numeric equivalent of the number we found so it can be used in caller
code that expects a value to be stored in "*a". But more importantly, STX_ERROR_WARN(ERR_NUMOFLOW)
is done. And IS_STX_WARN macro in compiler.h is modified to include ERR_NUMOFLOW as a parse-time
warning. With this a NUMOFLOW is treated just like an INVSVN would be treated in the middle of an
expression. It will not abandon the parse/triple-generation in the middle.

Additional changes include

* sr_port/toktyp.h and sr_port/mtables.c : A few filler entries were created before adding the
  new TK_NUMOFLOWLIT so YottaDB maintains compatibility with future GT.M changes in this area.

* sr_port/unary_tail.c and sr_port/ex_tail.c : Literal/Numeric optimizations are not done in case
  we encounter a numeric value with a NUMOFLOW. In sr_port/unary_tail.c, a NUMOFLOW stx_error is
  issued in one case when MV_FORCE_NUMD() does not return an mval with MV_NM mvtype set.

* sr_unix/rts_error.c : TREF(director_token) was set to TK_ERROR unconditionally. This seemed too
  arbitrary and so has been deleted. No test failures were seen because of this so it is considered
  a safe change i.e. no regressions at least.

* sr_port/mtables.c : Add literal_numoflow mval to store an impossible numeric value (exponent -= EXPHI).

* sr_port/mtables.c : Fix missing TK_SQLFIELDNAME definition in tokentable[]. This was a pre-existing
  uninitialized variable issue that showed up only in pro where NUMOFLOW error places were instead
  getting EXPR error.  Turns out tokentable[] array was missing a line entry for TK_SQLFIELDNAME all
  along.  Which was causing TK_CR and later token types (including the newly added TK_NUMOFLOWLIT)
  to be interpreted incorrectly (their actual index in the tokentable array would be one before). So
  accessing tokentable[TK_NUMOFLOWLIT] actually accessed one index after the actual tokentable[]
  index where TK_NUMOFLOWLIT was stored so we were looking at garbage (since TK_NUMOFLOWLIT is the
  last allocated entry in tokentable) and in pro tokentable[TREF(window_token)].uo_type turned out
  to be some garbage non-zero value (when TREF(window_token) was TK_NUMOFLOWLIT) and caused the flow
  to be different and incorrect.

* sr_port/mset.c : An unnecessary set of "s1" was removed.

* sr_port/expritem.c : Store literal_numoflow as result of NUMOFLOW operation. We might have as well
  stored a literal_null mval instead since an OC_RTERROR triple would have been generated at the
  point this value is going to be used in the triple chain. But just for consistency sake it would be
  better to treat the result of the numeric overflow evaluation to be an impossible number instead of 0
  (which is what literal_null evaluates to).

* sr_port/cdbg_dump.c & sr_port/cdbg_dump.h : Add debugger-callable functions to dump triples on
  &t_orig and TREF(curtchain) chains. Sometimes (e.g. f_select(), m_set()), we do setcurtchain() calls
  to switch triple generation to a different chain than &t_orig. So having just one function to invoke
  from the debugger and examine the triple chain originating from &t_orig (cdbg_dump_triple_all())
  is not considered enough. There are now 2 functions cdbg_dump_t_orig() and cdbg_dump_curtchain().
  The former replaces cdbg_dump_triple_all(). They examine the triple chains from &t_orig and
  TREF(curtchain) respectively.

* sr_port/unary_tail.c : An assert(TREF(rts_error_in_parse)) was negated because NUMOFLOW is now
  a IS_STX_WARN type error code. But later it was realized that an rts_error_csa(ERR_NUMOFLOW)
  will cause a NUMOFLOW error to be issued in direct mode (when "run_time" is 1 even though we are
  compiling M code entered at the direct mode prompt) even if that part of the expression will never
  be reached (for example : IF $SELECT(1:1,1:"1E47"&1) write "Hello "). Therefore the rts_error_csa()
  was replaced with stx_error() and associated TREF(rts_error_in_parse) usages were removed.

* sr_port/compiler.h, sr_port/ins_errtriple.c & sr_port/f_char.c : Fix INVDLRCVAL error just like the
  NUMOFLOW error was fixed above. By treating INVDLRCVAL as a parse warning (adding it to the list of
  allowed warning messages during parse in the IS_STX_WARN macro in sr_port/compiler.h). This also
  meant removing its use from sr_port/ins_errtriple.c which is rendered obsolete by the IS_STX_WARN
  addition. And returning a dummy result from sr_port/f_char.c in this case but ensuring the stx_error()
  is done so an error triple is inserted that will issue an INVDLRCVAL error at runtime if ever this
  codepath is reached in the execution.

Fix INVDLRCVAL error in $SELECT to behave just like NUMOFLOW error

That is $SELECT should stop evaluating tvexprs after the first true tvexpr is seen.
If later tvexpr/exprs have a INVDLRCVAL error they should not show up at runtime.