[#372] GTMASSERT2 fatal error when compiling SET command where first setleft is an invalid ISV and later setleft used indirection
Compiling M lines like the following used to terminate with a GTMASSERT2 error.

	set ($iv,@y(2))=1

This is because the moment an invalid ISV is detected in the first setleft in a compound SET,
we generate triples for the remainder of the SET command in a different/discarded chain.
But if indirection setlefts are encountered, we would invoke the INSERT_INDSAVGLVN macro which
would insert OC_GLVNSLOT, OC_INDSAVGLVN triples in the discard-chain. But at the very end,
we were inserting a OC_GLVNPOP triple if "used_glvn_slot" is TRUE. And this variable was true
if OC_GLVNSLOT triples were generated whether or not they got generated in the regular curtchain
or the discard chain. So if the OC_GLVNSLOT got generated in the discard chain we ended up with
a hanging OC_GLVNPOP in the main chain which later caused the GTMASSERT.

The fix is to ensure "used_glvn_slot" is set to TRUE only if we are in the main chain.