[#374] Strip any trailing CR from $TEXT return value if line is coming from embedded source
fill_src_tbl() invoked fill_src_tbl_via_litpool() in case the M source was embedded in the
object file and invoked fill_src_tbl_via_mfile() otherwise. The latter was already doing the
job of stripping <CR> but the former was not. This commit fixes the former too. That way
$TEXT output is consistent whichever of these 2 functions function it ends up calling.