Do not issue confusing GDECHECK message before the actual error (e.g. ENO13 Permission denied while writing to gld file at exit time)
If one does a EXIT at the GDE> prompt after changing some gld fields and the .gld file has read-only permissions,
one currently gets the following message.

%GDE-I-GDUPDATE, Updating Global Directory file
        /extra1/testarea1/nars/test/temp/tmp/tmp/mumps.gld
%GDE-I-GDECHECK, Internal GDE consistency check
%SYSTEM-E-ENO13, Permission denied

The GDECHECK message is confusing. The real issue is the ENO13 and seeing that right after the GDUPDATE is
enough for the user to know there was a permission error writing to the gld file. The GDECHECK message is
printed always before the actual error when we are about to abort from GDE. It seems unnecessary and could
be confusing like in the above example so it is now removed.

Below is the output one would see when one tries to exit (and write changes to gld) when gld is read-only.

%GDE-I-GDUPDATE, Updating Global Directory file
        /extra1/testarea1/nars/test/temp/tmp/tmp/mumps.gld
%SYSTEM-E-ENO13, Permission denied