[#349] MUPIP REORG on a database file with non-zero RESERVED_BYTES can cause integrity errors
The primary issue is in mu_reorg.c where the maximum space allowed in a block taking the input
fill factor as well as the current reserved_bytes setting in the db file header was not computed
correctly. The computation was subtracting reserved_bytes after doing the fillfactor% of the block size.
This could lead to negative values for d_max_fill and/or i_max_fill variables. That could in turn
cause out-of-design situations in caller functions mu_split() and in turn locate_block_split_point()
which could return with a split point that is before the first record in the block effectively
resulting in an empty left block after the split thus causing an integrity error (see r124/ydb349 subtest).

The computation is now corrected to subtract the reserved_bytes out of the block_size. This is
guaranteed to be a positive number. And then do the fillfactor% on this positive number. That
automatically fixes the issue.