[#191] Use GT.CM GNP server for @<hostname>:<filepath> syntax even if <hostname> is local host
When <hostname>:<filepath> is specified as the file name in the segment definition in the gld,
at runtime, YottaDB checks if <hostname> is a local host and if so does not go to GT.CM GNP.
Instead avoids using a GNP server, accesses the database locally and avoids network communication
overhead (aka short-circuiting). If not a local host, it goes to the GT.CM GNP server.

But with the @ prefix (i.e. @<hostname>:<filepath>), YottaDB now goes to the GT.CM GNP server
even if <hostname> is a local host.