[#361] Fix rare manually_start/4g_journal subtest failure (GTMASSERT2 in update process)
See https://gitlab.com/YottaDB/DB/YDB/issues/361 for original details of #361.

The manually_start/4g_journal subtest is what failed a dbg-only assert and caused #361
to be created in the first place. A fix was put in but it was incomplete as evidenced by
another rare failure of the same test this time in a pro build.

This time, the update process failed an assertpro (i.e. a fatal error).

%YDB-F-GTMASSERT2, YottaDB r999 Linux x86_64 - Assert failed sr_port/updproc.c line 1428
	for expression (FALSE && (upd_proc_local->read_jnl_seqno == jnlpool_ctl_seqno))

In the failure case, upd_proc_local->read_jnl_seqno was 2 but jnlpool_ctl_seqno was 1.  This failed
after having played 1 seqno. In the test run, it was an A->P setup. That is A was a non-supplementary
instance replicating to P a supplementary instance. The strm_index was chosen to be 1. That is,
all seqnos on A will be mapped to a 60-bit strm_seqno on P where the most significant 4-bits of
the 64-bit seqno would be 0001 (binary) corresponding to a strm_index of 1.

Since P is a supplementary instance, jnlpool_ctl_seqno was derived from the variable
jnlpool->jnlpool_ctl->strm_seqno[strm_index]. And since this was 1 after playing one seqno it means
it was 0 before any seqno was played. But that is puzzling since starting the receiver server with
-updateresync should have set this to 1 as part of writing the new history record the first time
A->P connection happens. That did not happen in this test.

Turns out A->P connection happened, but before the history record could be sent by the source server
on A, the receiver server on P was shut down by the test and later when the receiver server was
restarted, the test framework started it without the -updateresync qualifier which meant it did not
set the HISTINFO_TYPE_UPDRESYNC history_type in the modified history record stored in the receive
pool. This caused the update process to not set jnlpool->jnlpool_ctl->strm_seqno[] to a non-zero
value which it would have done for the HISTINFO_TYPE_UPDRESYNC case. That is the cause.

The fix is in the second receiver startup to derive the fact that the first receiver startup was a
-updateresync one. This is done by checking if the initial handshake went okay and if there is no
history record ever written for the "strm_index" stream. If so, that means the first receiver startup
should have been a -updateresync one and so assume that.

Note that this GTMASSERT2 error is only a post-r1.22 issue due to the prior #361 commit being an
incomplete fix.  r1.22 and prior versions would fail with a REPLINSTNOHIST error in this case.

The manually_start/4g_journal subtest exercises this codepath only rarely (once a month) but the
recently written r124/ydb361 subtest does fail once in 10 runs or so on the ARMV7L boxes with the
same GTMASSERT2 error if it is enhanced to do a few updates. So that is considered a good automated
test of this issue (very unlikely in practice since -updateresync startups are not immediately shut
down before the first seqno gets replicated).