Fix incorrect assert (rare debug-only v53002/C9E04002596 subtest failure)
In one run of the v53002/C9E04002596 subtest, the following assert failed.

%YDB-F-ASSERT, Assert failed in sr_port/mutex_deadlock_check.c line 102 for expression
	(!jnlpool || !jnlpool->jnlpool_dummy_reg || jnlpool->jnlpool_dummy_reg->open || (repl_csa->critical != criticalPtr) || (NULL == cs_addrs))

And below was the C-stack at the time of the assert failure.

(gdb) where
 #0  kill () at ../sysdeps/unix/syscall-template.S:84
 #1  gtm_dump_core () at sr_unix/gtm_dump_core.c:69
 #2  gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:148
 #3  ch_cond_core () at sr_unix/ch_cond_core.c:64
 #4  rts_error_va () at sr_unix/rts_error.c:159
 #5  rts_error_csa () at sr_unix/rts_error.c:92
 #6  mutex_deadlock_check () at sr_port/mutex_deadlock_check.c:101
 #7  mutex_long_sleep () at sr_unix/mutex.c:511
 #8  gtm_mutex_lock () at sr_unix/mutex.c:856
 #9  grab_lock () at sr_unix/grab_lock.c:83
 #10 repl_inst_ftok_counter_halted () at sr_unix/repl_inst_ftok_counter_halted.c:45
 #11 jnlpool_init () at sr_unix/jnlpool_init.c:764
 #12 gvcst_init () at sr_port/gvcst_init.c:917
 #13 gv_init_reg () at sr_port/gv_init_reg.c:56
 #14 gv_bind_name () at sr_port/gv_bind_name.c:75
 #15 op_gvname_common () at sr_port/op_gvname.c:117
 #16 op_gvname_fast () at sr_port/op_gvname.c:81

The assert at line 102 expects that if ever we are in mutex_deadlock_check() for the jnlpool, we better have
not opened any database file (i.e. non-NULL cs_addrs). Whereas in this call sequence clearly, we have a case
where while opening the database, we need to open the journal pool (because ydb_custom_errors env var is set)
and while opening the journal pool we notice a 32K semaphore counter overflow (artificially created by the test
system in this debug-run using ydb_db_counter_sem_incr env var) which results in getting a lock on the jnlpool
before making changes to the instance file and while trying to get the lock we notice it is being held by
some other process which is taking a long time so we go to mutex_deadlock_check() eventually for the jnlpool
and fail the assert while in the middle of a database open (so cs_addrs is non-NULL).

The assert is clearly wrong for this case.

The assert was only introduced in V6.3-001A in the below state.

assert((NULL == jnlpool.jnlpool_dummy_reg) || jnlpool.jnlpool_dummy_reg->open || (repl_csa->critical != criticalPtr));

And it was modified in V6.3-003 and V6.3-005 to add || conditions. Most likely to account for exceptions to
the assert as they were encountered.

I don't see the value in this assert so removing it.

Also corrected a pre-existing comment a few lines after the removed assert to reflect our renewed understanding
of the current failure possibility.