[#313] MUPIP FTOK -JNLPOOL and -RECVPOOL should not ignore input replication instance file name
repl_inst_get_name() was called unconditionally before and that set the "instfilename" variable
to point to the instance file name pointed to by the "ydb_repl_instance" env var. And ignored
the filename specified by the user in the command line ("fn"). This call to repl_inst_get_name()
is now removed and the repl_inst_read() is done with "fn" instead of "instfilename". That
ensures the user specified file name is used in MUPIP FTOK -JNLPOOL and MUPIP FTOK -RECVPOOL.