[#96] Record DSE commands that change database blocks in the AIMG (after-image) journal record; MUPIP JOURNAL -EXTRACT -DETAIL extracts this out in user-readable form
There was already a 4-byte filler in the struct_jrec_blk structure that is shared by the PBLK and
AIMG records. This filler is now used to store the length of the DSE command that is stored at the
tail end of the record (i.e. after the after-image block contents) in case of an AIMG record (new
"cmdstrlen" field). The "cmdstrlen" field is unused in case of PBLK records.

COMPUTE_AIMG_CHECKSUM macro was enhanced to do what COMPUTE_PBLK_CHECKSUM was previously doing and
additionally compute the checksum of the DSE command line too and add it to the AIMG record checksum.

Since the maximum block size is 64K, the maximum AIMG record size now rose from approximately 64K
to 64K + maximum command line length (i.e. MAX_LINE which is 32K) = 96K. Since this is a lot less
than the maximum journal record length (which is approximately 1Mb), there are no issues with the
modified AIMG record crossing any existing limits (and in turn posing problems).

The DSE command line is stored as a null-terminated string in cli_lex_in_ptr->in_str. There is no field
currently that indicates the length of this string (cli_lex_in_ptr->buflen stores the allocated length
of cli_lex_in_ptr->in_str, not the used length). This means we need to do a STRLEN in t_end.c in case
of the AIMG record. But since all this overhead is only in the AIMG record codepath (which is not the
common case), it is considered okay for now.

Since the dse command line is added after the block contents in the AIMG journal record, there is no
change to mur_put_aimg_rec() since it currently only looks at the block contents part of the AIMG record.

Since the AIMG record format is changing, the journal record format needs to be bumped but since
it was bumped only recently (as part of #362, the NULL record format changed), there is no need to
do one more bump since both #362 and #96 are going to be released in r1.24.