[#376] Add support for TLS 1.3 in the reference TLS implementation that ships with the encryption plugin; [#375] Remove support for SSL2.0, SSL3.0 and TLS1.0 even for encrypted SOCKET devices using TLS plugin
Fixes for #376
---------------
The primary fix is to sr_unix/gtm_tls_impl.c to allow TLS 1.3 (which is the default ssl version in the
libssl.so library on Ubuntu 18.10) as a recognized version in the plugin.

While at this, removed openssl-1.0 code in the makefile (sr_unix/Makefile.mk) that was introduced a while
back. This was an Arch-Linux-only workaround introduced around a year back to force openssl1.0 to be linked
by the TLS plugin and is no longer needed now that we support openssl1.1.

With these changes, testing in-house revealed a non-trivial issue. TLS 1.3 does not support renegotiation
whereas that is a user-visible aspect of the TLS plugin. So SSL_renegotiate() calls were done only for
TLS 1.2 and below whereas in TLS 1.3 a SSL_key_update() call is now done to update the connection keys
which is the primary purpose (in my understanding) why the SSL_renegotiate() call is currently being
used in the plugin. This keeps the user-level view of a key renegotiation still in place even if TLS 1.3
is being used inside the plugin.

It was also noticed that assert() macro was enabled even in PRO builds of the TLS plugin (gtm_tls_impl.c)
whereas in the YottaDB code it is enabled only in DEBUG builds. This can have even user-visible impact
in cases where the assert() call invokes other functions. This macro is now fixed to be enabled only in
DEBUG builds.

SSL_get_verify_result() was being invoked unnecessarily (in my opinion) in gtm_tls_connect(),
gtm_tls_accept(), gtm_tls_send() and gtm_tls_recv(). Those calls are now removed and just SSL_get_error()
is invoked in case of an error.

Fixes for #375
---------------
The place where list of supported TLS versions was checked is not correct. It was in
gtm_tls_get_conn_info() in the TLS plugin which is called by the replication servers in TLS mode,
but not necessarily by encrypted SOCKET devices (in M) which acts as a server or client with TLS
enabled (and uses the TLS plugin). This check is now done in a new function ydb_tls_is_supported()
which is invoked from gtm_tls_connect() and gtm_tls_accept() (the client and server side
invocations) which are called both by the replication servers and encrypted SOCKET devices which
uses TLS functionality. This new function returns -1 in case of an unsupported TLS version (of the
negotiated connection). Additionally, it sets the error string to indicate the current TLS version
(e.g. "TLSv1.1") thus giving the user a more friendly error message in case the connection is not
supported by the TLS plugin.