[#369] Fix SETENVFAIL/UNSETENVFAIL error to use non-freed memory for env var name
This is a bug introduced in the prior #369 commit that was identified by the currently developed
r124/ydb369 subtest.

We were doing a free(envvarname) and then using it in the rts_error() call.
We need to do the free() before the rts_error() as control gets transferred elsewhere after that.
Therefore we instead use the env var name stored in arg->str (which is not a null-terminated string).
That means the SETENVFAIL error will now be passed a !AD instead of a !AZ and therefore the faocount
needs to also be bumped (changes to ydberrors.msg and ydberrors_ctl.c). Also, arg was being used
to store the env var name and later to store the env var value but since we now need the env var name
we need another variable "arg2" which stores the env var value.

Similar changes were needed in UNSETENVFAIL error.