[#365] Allow BEFORE or NOBEFORE to not be specified in MUPIP SET JOURNAL; Use prior journal type from db file header in that case
The primary fix is to sr_unix/mupip_cmd_disallow.c where any MUPIP SET JOURNAL command which had
BEFORE_IMAGE or NOBEFORE_IMAGE not specified was disallowed. That is now removed.

A related assert in sr_port/mupip_set_journal_parse.c (which relied on the above disallowing code)
is also removed.

The changes to sr_port/mupip_set_journal_newstate.c are cosmetic (to remove VMS_ONLY/UNIX_ONLY usages).

The changes to sr_port/jnl.h are cosmetic (to fix a typo).

Using prior journal type from db file header (in case BEFORE or NOBEFORE was not specified) was already
being done in sr_port/mupip_set_journal_newstate.c (when rptr->before_images is set to "current_image")
so no general code changes were needed related to this requirement.

But one subtle scenario that needed to be handled was the case where -JOURNAL=ENABLE is specified AND
the current journal state is disallowed AND the current access method is BG. In this case, normally
the value of csd->jnl_before_image would be TRUE (inherited from the gld which sets BEFORE_IMAGE as
the default for BG). But it is possible that one changes the access method to MM and switches it back
to BG using MUPIP SET commands after a MUPIP SET -JOURNAL=DISABLE command (which sets csd->jnl_before_image
to FALSE). In that case, csd->jnl_before_image would be FALSE AND access method is BG. If one now attempts
a MUPIP SET -JOURNAL=ENABLE,ON command, inheriting the before image state from the db file header would
cause a NOBEFORE_IMAGE journal file to be created which is incorrect. In this case, we need a BEFORE_IMAGE
type journal file as the default given the access method is BG and journaling state is changing from
DISALLOWED to ENABLED as part of the same command. This is fixed in sr_port/mupip_set_journal_newstate.c.