[#365] MUPIP SET -ACCESS_METHOD=BG should set default journaling type to BEFORE_IMAGE in db file header if access method is changing
This fix was necessitated by the following use-case.

	1) Create gld and database file with BG access method.
	2) mupip set -journal="enable,on" -reg "*"
	3) mupip set -journal=off -reg "*"
	4) mupip set -acc=MM -reg "*"
	5) mupip set -acc=BG -reg "*"
	6) mupip set -journal=on -reg "*"

If steps 1,2,3 and 6 are run, a journal file with BEFORE_IMAGE journaling is created.

But if steps 4,5 (which together should be a no-op since the access method is changing to MM and
being restored to BG again) are introduced in between steps 3 and 6, a journal file with
NOBEFORE_IMAGE journaling is created. This is incorrect.

The cause of the issue is in step (4) where switching access method to MM resets the db file header
csd->jnl_before_image to FALSE (i.e. NOBEFORE_IMAGE journaling).

This is now fixed to create journal file with BEFORE_IMAGE journaling. The fix is in mupip_set_file.c
so when the access method is switching to BG and journaling is currently not on in the database,
we set the default journaling type to BEFORE_IMAGE. So step (5) undoes the change due to step (4).