[AARCH64] [#358] Enable LIBAIO on 64-bit and 32-bit ARM builds of YottaDB
The r124/ydb358 subtest (which was enabled for 64-bit boxes) failed an assert in wcs_flu.c while
testing on the 64-bit ARM.

This test has a C program using simpleapi to do sets of global nodes in a parent process and a child
process after a fork()/ydb_child_init() sequence in the same C program. The whole experiment is done
while asyncio is enabled in the database. This is to test a fix to ydb_child_init() that went in
https://gitlab.com/YottaDB/DB/YDB/merge_requests/362 (as part of #358).

The assert failed in wcs_flu() because it timed out trying to flush the WIP (writes-in-progress) queue
of global buffers in the child process in the C program.  The aiocbp (AIO control block) in the
cache-record in the WIP queue always held a status of EINPROGRESS. No separate user-level thread
was started to do the aio_write() in the child like one normally sees when the parent does an aio_write().

After a while looking for potential issues, I could find none. That is when it struck me that the
POSIX AIO implementation (aio_write, aio_error, aio_return, aio_cancel etc.) could have a limitation
in that it does not work in case a POSIX AIO write happened in a parent process and a fork occurs
(without an exec) and then a child process does a POSIX AIO write too (just like is done in this test).

This is because global variables in the parent that are related to maintaining the POSIX AIO state
are inherited by the child. And there is no POSIX AIO function I could find to clear this state in
the child (in YottaDB, we have ydb_child_init() exactly for that purpose). And so I don't think
simpleAPI + forks + asyncio will work on ARM 32-bit or 64-bit boxes.

On the other hand with LIBAIO (io_submit() etc.) there are no such issues. This ydb358 subtest
works fine on x86_64 which have LIBAIO currently enabled.

Given this finding, it seems better to enable LIBAIO (instead of POSIX AIO) on the ARM boxes
(both 32-bit and 64-bit). That should let simpleAPI + forks + asyncio work on 32-bit/64-bit ARM too.